+++ examples/main.c
@@ -175,7 +175,7 @@ int main(int argc, char *argv[]) {
   char *cpufreqarg = NULL;
   char *param_filename = NULL;
   char restart_file[200] = "";
-  unsigned long long cpufreq = 0;
+  uint64_t cpufreq = 0;
   struct cmdparams cmdps;
   cmdps.nparam = 0;
   cmdps.param[0] = NULL;
@@ -346,7 +346,7 @@ int main(int argc, char *argv[]) {
   }
 #endif
 
-  /* The CPU frequency is a long long, so we need to parse that ourselves. */
+  /* The CPU frequency is a int64_t, so we need to parse that ourselves. */
   if (cpufreqarg != NULL) {
     if (sscanf(cpufreqarg, "%llu", &cpufreq) != 1) {
       if (myrank == 0)
@@ -941,10 +941,10 @@ int main(int argc, char *argv[]) {
 #endif
 
     /* Get the total number of particles across all nodes. */
-    long long N_total[swift_type_count + 1] = {0};
-    long long Nbaryons = Ngas + Nspart + Nbpart;
+    int64_t N_total[swift_type_count + 1] = {0};
+    int64_t Nbaryons = Ngas + Nspart + Nbpart;
 #if defined(WITH_MPI)
-    long long N_long[swift_type_count + 1] = {0};
+    int64_t N_long[swift_type_count + 1] = {0};
     N_long[swift_type_gas] = Ngas;
     N_long[swift_type_dark_matter] =
         with_gravity ? Ngpart - Ngpart_background - Nbaryons : 0;
@@ -1137,7 +1137,7 @@ int main(int argc, char *argv[]) {
 
     /* Get some info to the user. */
     if (myrank == 0) {
-      const long long N_DM = N_total[swift_type_dark_matter] +
+      const int64_t N_DM = N_total[swift_type_dark_matter] +
                              N_total[swift_type_dark_matter_background];
       message(
           "Running on %lld gas particles, %lld stars particles %lld black "
